thermalComfort/requirements.txt
django==1.11
gunicorn
pillow
psycopg2==2.7.3.2
djangorestframework
django-rest-swagger
static3==0.5.1
django-storages
boto==2.42.0
django_cors_headers==2.1.0
validate_email==1.3
boto3==1.5.27
pymongo==3.6.0
geoip2==2.7.0
django-crum==0.7.2
django-import-export==1.0.0
